Feat: Send confirtm mail payment subscription
