--- api/src/utils/db.ts
@@ -1,14 +1,0 @@
-import { PrismaClient } from "@prisma/client";
-
-let db: PrismaClient;
-
-if (process.env.NODE_ENV === "production") {
-  db = new PrismaClient();
-} else {
-  // Ensure the prisma instance is re-used during hot-reloading
-  // Otherwise, a new client will be created on every reload
-  globalThis["db"] = globalThis["db"] || new PrismaClient();
-  db = globalThis["db"];
-}
-
-export default db;
