+++ client/.env-sample
@@ -2,4 +2,5 @@ VITE_API_URL_BASE="https://empli-www-production.up.railway.app/api"
 VITE_CLOUD_PRESET="intikfxk"
 VITE_CLOUD_URL="https://api.cloudinary.com/v1_1/dvzhqzjkm/image/upload"
 VITE_CLERK_PUB_KEY="pk_test_ZGVsaWNhdGUtY29kLTg2LmNsZXJrLmFjY291bnRzLmRldiQ"
-VITE_MP_PUB_KEY=
+VITE_MP_PUBLIC_KEY= "TEST-575ad886-b1ec-4b11-bd3a-eddd47b135ff"
+VITE_EMAILJS_PUB_KEY="NSBq0dEL_bYCIyvJs"
