+++ client/src/components/Cards/CardLarge/CardLarge.jsx
@@ -1,41 +1,38 @@
 /* eslint-disable prettier/prettier */
 import { PlusIcons } from 'components'
-// import {
-//   useGetJobsQuery,
-//   useGetAllTalentsQuery,
-// } from '@/redux/features/api/base'
+import {
+  useGetJobsQuery,
+  useGetAllTalentsQuery,
+} from '@/redux/features/api/base'
 
 export const CardLarge = () => {
   // info sirve unicamente para probar que la card renderiza,
   // porque lo que se quiere es que la info venga por props, asi todo seria dinamico
-  // const { data: offersData = [] } = useGetJobsQuery()
-  // const { data: talentsData = [] } = useGetAllTalentsQuery()
+  const { data: offersData = [] } = useGetJobsQuery()
+  const { data: talentsData = [] } = useGetAllTalentsQuery()
 
   // total talentsData in the aplication. it is array
-  // const countTalents = talentsData && talentsData.length
-  // console.log(countTalents)
+  const countTalents = talentsData && talentsData.length
 
-  //   offersData &&
-  //   offersData.reduce((acu, curr) => {
-  //     const name = curr.company.name
-  //     if (!acu[name]) {
-  //       acu[name] = true
-  //     }
-  //     return acu
-  //   }, {})
+  const uniqueCompany = offersData &&
+    offersData.reduce((acu, curr) => {
+      const name = curr.company.name
+      if (!acu[name]) {
+        acu[name] = true
+      }
+      return acu
+    }, {})
 
-  // const countCompany = Object.keys(uniqueCompany).length
-  // console.log(countCompany)
+  const countCompany = Object.keys(uniqueCompany).length
 
   // total offersData in the aplication. it is array
-  // const countOffers = offersData && offersData.length
-  // console.log(countOffers)
+  const countOffers = offersData && offersData.length
+  console.log(countOffers)
   const Marked = ({ data, title, isBorder = false }) => {
     return (
       <div
-        className={`flex w-1/3 flex-col items-center text-center transition ${
-          isBorder ? 'border-gray border-l-2 border-r-2' : ''
-        } duration-300 hover:scale-125 md:text-left`}
+        className={`flex w-1/3 flex-col items-center text-center transition ${isBorder ? 'border-gray border-l-2 border-r-2' : ''
+          } duration-300 hover:scale-125 md:text-left`}
       >
         <h1 className="mb-3 text-center font-quicksand text-xs font-bold uppercase tracking-wider text-gray-400">
           {title}
@@ -52,11 +49,11 @@ export const CardLarge = () => {
   }
   return (
     <div className="relative mx-auto flex w-full flex-col rounded-3xl bg-white shadow-2xl md:flex-row md:items-center md:justify-center md:p-2 lg:p-10">
-      <Marked data={'10k'} title="Profesionales" />
+      <Marked data={countTalents} title="Profesionales" />
 
-      <Marked data={'3k'} title="Ofertas" isBorder={true} />
+      <Marked data={countOffers} title="Ofertas" isBorder={true} />
 
-      <Marked data={'5k'} title="Empresas" />
+      <Marked data={countCompany} title="Empresas" />
     </div>
   )
 }
