+++ client/src/components/Forms/UserForms/inputCls.js
@@ -1,0 +1,2 @@
+
+export const inputCls = 'rounded-md border border-slate-200 p-3 outline-none focus:ring-2';
