+++ client/src/pages/index.js
