+++ client/src/utils/middleware/paymentId.js
@@ -1,0 +1,3 @@
+export const paymentId = ()=>{
+    
+}
